+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
-## Unreleased
+## [1.2.0] - 2016-04-22
 ### Added
 * `stream.abort()` will kill ogr and cleanup temp files
 
@@ -120,7 +120,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 * Initial release: API supports creating VRTs, Geojson, shapefile, kml and csv
 
-[Unreleased]: https://github.com/koopjs/geoxform/compare/v1.1.5...HEAD
+[Unreleased]: https://github.com/koopjs/geoxform/compare/v1.2.0...HEAD
+[1.2.0]: https://github.com/koopjs/geoxform/compare/v1.1.5...v1.2.0
 [1.1.5]: https://github.com/koopjs/geoxform/compare/v1.1.4...v1.1.5
 [1.1.4]: https://github.com/koopjs/geoxform/compare/v1.1.3...v1.1.4
 [1.1.3]: https://github.com/koopjs/geoxform/compare/v1.1.2...v1.1.3